# Changelog for week 44/45

**Tagion HEAD record**
The tagion HEAD name record stores all the global statistics about the current network in the DART.
Like the epochnumber, total money supply, number of bill, etc..

**Genesis Epoch added**
The Genesis Epoch is the network boot record.
Where we the system stores all the information required to bootstrap the network.

**Hashgraph epoch number**
In order to stay clear of future overflows,
the hashgraph epochnumber and the transcript service has been switchted from an int (32-bit) to a long (64-bit).

**Epoch Votes**
The epoch is now created based on the votes of the DART bullseye.

**MuSig2**
The NativeSecp256k1 module now features Schnorr Multi Signatures and has been unittested.

**Fixes & Stabillity improvements**
 * We have made changes to how the node starts the replicator service. 
 * Improved the way the transcript stores and cleans votes.
 * Removed some unsafe type casts.
 * Fixed error with unflushed DART writes.
 * Created improvements to the Error replys when sending a contract and making a DART read request.
 * Fixed the inputvalidator test, which used the wrong socket type.
 * ...

**Remove legacy code**
We have removed all of the SSL modules.

**CI Improvements**
The CI flow now runs in several steps, so we have better error reporting when and which stage fails.
The worflows times out when a job hangs. And it always produces an artifact so we can inspect the errors

+++ src/lib-hashgraphview/tagion/hashgraphview/Compare.d
@@ -26,7 +26,7 @@ struct Compare {
     const HashGraph h1, h2;
     const ErrorCallback error_callback;
     int order_offset;
-    int round_offset;
+    long round_offset;
     uint count;
     this(const HashGraph h1, const HashGraph h2, const ErrorCallback error_callback) {
         this.h1 = h1;
